+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Fix adapter codegen to avoid memory shift overflow, and support SIMD instructions. ([#629](https://github.com/fastly/Viceroy/pull/629))
 - Add support for sending `PendingRequestHandle` downstream ([#631](https://github.com/fastly/Viceroy/pull/631))
+- Update to the latest WITs and adapter. ([#637](https://github.com/fastly/Viceroy/pull/637))
 - Fix to correctly apply TLS when using `wss:` URL schema with WebSockets passthrough ([#636](https://github.com/fastly/Viceroy/pull/636))
 
 ## 0.18.0 (2026-05-21)